authorization implemented

+++ src/api/notes/handler.js
@@ -16,8 +16,9 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload)
       const { title = 'untitled', body, tags } = request.payload
+      const { id: credentialId } = request.auth.credentials
 
-      const noteId = await this._service.addNote({ title, body, tags })
+      const noteId = await this._service.addNote({ title, body, tags, owner: credentialId })
 
       const response = h.response({
         status: 'success',
@@ -52,7 +53,8 @@ class NotesHandler {
   }
 
   async getAllNotesHandler(request, h) {
-    const notes = await this._service.getAllNotes()
+    const { id: credentialId } = request.auth.credentials
+    const notes = await this._service.getAllNotes(credentialId)
 
     const response = h.response({
       status: 'success',
@@ -67,6 +69,9 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params
+      const { id: credentialId } = request.auth.credentials
+
+      await this._service.verifyNoteOwner(id, credentialId)
       const note = await this._service.getNoteById(id)
 
       return {
@@ -101,6 +106,9 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload)
       const { id } = request.params
+      const { id: credentialId } = request.auth.credentials
+
+      await this._service.verifyNoteOwner(id, credentialId)
       await this._service.editNoteById(id, request.payload)
 
       return {
@@ -132,6 +140,9 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params
+      const { id: credentialId } = request.auth.credentials
+
+      await this._service.verifyNoteOwner(id, credentialId)
       await this._service.deleteNoteById(id)
 
       return {
